requirements.txt
# pip default config
--index-url https://pypi.python.org/simple/
pyqt>=5.6